[OWL-1866] Move MysqlApiService to service package.
