+++ modules/hbs/http/health.go
@@ -5,11 +5,12 @@ import (
 
 	"github.com/Cepave/open-falcon-backend/modules/hbs/g"
 	"github.com/Cepave/open-falcon-backend/modules/hbs/rpc"
+	"github.com/Cepave/open-falcon-backend/modules/hbs/service"
 	"github.com/gin-gonic/gin"
 )
 
 func getHealth(c *gin.Context) {
-	mysqlInfo := MysqlApiService.GetHealth()
+	mysqlInfo := service.MysqlApiService.GetHealth()
 	httpInfo := g.Config().Http
 	rpcInfo := &g.RpcView{g.Config().Listen}
 	heartbeatService := rpc.AgentHeartbeatService
